refactor(desktop): avoid project sort module overhead
Keep project-tree preference helpers in the existing topic utility module so the extracted code satisfies both source-size and initial-bundle budgets.
